+++ sdk/identity/identity/test/internal/node/defaultAzureCredential.spec.ts
@@ -125,7 +125,12 @@ describe("create functions", () => {
     expect(psSpy).not.toHaveBeenCalled();
 
     expect(miSpy).toHaveBeenCalledWith({
-      disableProbe: true,
+      allowInsecureConnection: true,
+      retryOptions: {
+        maxRetries: 5,
+        retryDelayInMs: 800,
+      },
+      sendProbeRequest: false,
     });
   });